fixup! Circuit padding header.
Switch circpad_hist_token_t to uint32_t. uint16_t is only 32mb of data per
bin, which may not be enough if a histogram wants to encode eg a large page
fetch in bin 0.

+++ src/core/or/circuitpadding.h
@@ -62,7 +62,7 @@ typedef enum {
 } circpad_decision_t;
 
 /** The type for the things in histogram bins (aka tokens) */
-typedef uint16_t circpad_hist_token_t;
+typedef uint32_t circpad_hist_token_t;
 
 /** The type for histogram indexes (needs to be negative for errors) */
 typedef int8_t circpad_hist_index_t;
